extract SubjectsRow compoenent
